+++ resources/views/hr/salary-set/_index_filtered.blade.php
@@ -14,7 +14,7 @@
 </div>
 
 <div class="table-body-wrapper">
-    @foreach($settingsSalarySets as $key=>$item)
+    @forelse($settingsSalarySets as $key=>$item)
         <div class="table-body-item-wrapper d-flex flex-wrap">
             <div class="table-body-item dep-list">
                 <h4>{{ $settingsSalarySets->firstItem() + $loop->iteration - 1 }}</h4>
@@ -27,20 +27,28 @@
                 <h4 class="text-center erp-t-email"> {!! $item->description ?? 'N/A' !!} </h4>
             </div>
             <div class="table-body-item dep-list pe-2 justify-content-end">
-                <div class="erp-action-t">
-                    <div class="dropdown dropdown-action">
-                        <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
-                        <div class="dropdown-menu dropdown-menu-right">
-                            <a class="dropdown-item" href="javascript:void(0)" onclick="editItem({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a>
-                            <a class="dropdown-item" href="{{ route('hr.salary-set.set-employees',$item->id) }}" ><i class="fa-solid fa-pencil m-r-5"></i> Set Employees</a>
-                            <a class="dropdown-item" href="javascript:void(0)" onclick="setLeaveTypes({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Set Leave Types</a>
-                            <a class="dropdown-item" href="javascript:void(0)" onclick="setAttendance({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Set Attendance</a>
-                            <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('hr.salary-set.delete',$item->id) }}', 'reloadAjaxGetData')"><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
+                @if(hasPermission('manage-salary-set'))
+                    <div class="erp-action-t">
+                        <div class="dropdown dropdown-action">
+                            <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
+                            <div class="dropdown-menu dropdown-menu-right">
+                                <a class="dropdown-item" href="javascript:void(0)" onclick="editItem({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a>
+                                <a class="dropdown-item" href="{{ route('hr.salary-set.set-employees',$item->id) }}" ><i class="fa-solid fa-pencil m-r-5"></i> Set Employees</a>
+                                <a class="dropdown-item" href="javascript:void(0)" onclick="setLeaveTypes({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Set Leave Types</a>
+                                <a class="dropdown-item" href="javascript:void(0)" onclick="setAttendance({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Set Attendance</a>
+                                <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('hr.salary-set.delete',$item->id) }}', 'reloadAjaxGetData')"><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
+                            </div>
                         </div>
                     </div>
-                </div>
+                @endif
             </div>
         </div>
-    @endforeach
+    @empty
+        <div class="table-body-item-wrapper">
+            <div class="table-body-item text-center d-block text-primary">
+                No Data Found..!
+            </div>
+        </div>
+    @endforelse
 </div>
 {{ $settingsSalarySets->links('vendor.pagination.common_ajax_pagination') }}
