+++ resources/views/hr/user-resignation/_index_filtered.blade.php
@@ -10,11 +10,13 @@
                 <th class="erp-th text-center">Notice Date </th>
                 <th class="erp-th text-center">Resignation Date </th>
                 <th class="erp-th text-center">Status </th>
-                <th class="text-end erp-th">Action</th>
+                @if(hasPermission('manage-employee-resignation'))
+                    <th class="text-end erp-th">Action</th>
+                @endif
             </tr>
         </thead>
         <tbody class="erp-tbody">
-            @foreach($userResignations as $item)
+            @forelse($userResignations as $item)
                 <tr class="erp-tbody-tr">
                     <td class="erp-tbody-td">
                         <h4 class="d-table-title">1</h4>
@@ -61,23 +63,30 @@
                             </div>
                         </div>
                     </td>
-                    <td class="text-end erp-tbody-td">
-                        <div class="erp-action-t">
-                            <div class="dropdown dropdown-action">
-                                <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
-                                <div class="dropdown-menu dropdown-menu-right">
+                    @if(hasPermission('manage-employee-resignation'))
+                        <td class="text-end erp-tbody-td">
+                            <div class="erp-action-t">
+                                <div class="dropdown dropdown-action">
+                                    <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
+                                    <div class="dropdown-menu dropdown-menu-right">
+                                        <a class="dropdown-item" href="javascript:void(0)" onclick="editItem({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a>
+                                        <a class="dropdown-item" href="javascript:void(0)" onclick="updateStatus(this, function () { getData() })" data-href="{{ route('hr.user-resignation.change-status',[$item->id,1]) }}"><i class="fa-solid fa-check m-r-5"></i> Approve</a>
+                                        <a class="dropdown-item" href="javascript:void(0)" onclick="rejectItem({{$item->id}})"><i class="fa-solid fa-ban m-r-5"></i> Reject</a>
+                                        <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('hr.user-resignation.delete',$item->id) }}', 'reloadAjaxGetData')"><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
 
-                                    <a class="dropdown-item" href="javascript:void(0)" onclick="editItem({{$item->id}})"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a>
-                                    <a class="dropdown-item" href="javascript:void(0)" onclick="updateStatus(this, function () { getData() })" data-href="{{ route('hr.user-resignation.change-status',[$item->id,1]) }}"><i class="fa-solid fa-check m-r-5"></i> Approve</a>
-                                    <a class="dropdown-item" href="javascript:void(0)" onclick="rejectItem({{$item->id}})"><i class="fa-solid fa-ban m-r-5"></i> Reject</a>
-                                    <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('hr.user-resignation.delete',$item->id) }}', 'reloadAjaxGetData')"><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
-
+                                    </div>
                                 </div>
                             </div>
-                        </div>
+                        </td>
+                    @endif
+                </tr>
+            @empty
+                <tr class="erp-tbody-tr">
+                    <td class="erp-tbody-td text-center text-primary" colspan="9">
+                        Data not found..!
                     </td>
                 </tr>
-            @endforeach
+            @endforelse
         </tbody>
     </table>
 </div>
